Add test coverage for the web app
Additionally, configure a webpack build option for creating test
builds. This way when we are making our distribution build, we will get
a build created in the dist folder with app/src/page.js as our entry
point, while if we make our test build, the build is created in the
test-dist folder with the test files in app/test as our entry point.